Possivel realizar a operações CRD do CRUD na pagina de compras

+++ src/main/java/com/gobra/sistemagestaodeobras/dto/CompraRequestDTO.java
@@ -21,10 +21,10 @@ public record CompraRequestDTO (
 
   Date dataVencimento,
 
-  Integer valorOriginal,
+  double valorOriginal,
 
-  Integer valorDesconto,
+  double valorDesconto,
 
-  Integer valorFinal
+  double valorFinal
 
 ) { }
